+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Change Log
 
+## [0.0.4]
+
+- Added "Log retention time" feature
+- Added "Delete excluded folders" feature
+- Added "Delete empty folders" feature
+
 ## [0.0.3]
 
-- Added possibility to set folder to exclude 
+- Added "Exclude folder" feature 
 
 ## [0.0.2]
 
-- Bump version to 0.0.2
-- Set minimum VS Code version
+- Set minimum VS Code version to 1.66.1
 - Fix file age not visible if file has long name
 - Fix error when opening log before opening working folder